small refactor + ignore inline tags when matching sections with documents
