+++ arretify/step_segmentation/core.py
@@ -17,259 +17,25 @@
 # limitations under the License.
 #
 from typing import (
-    Callable,
     List,
     TypeGuard,
     Dict,
     Any,
-    Tuple,
-    Iterator,
     Union,
-    TypeVar,
     cast,
-    Generic,
 )
 from dataclasses import dataclass, field
 
 from arretify.types import TextSegment
-from arretify.utils.functional import iter_func_to_list
 from arretify.regex_utils import PatternProxy, regex_tree
 from arretify.regex_utils.regex_tree.execute import match
+from arretify.utils.split_merge import (
+    make_while_splitter,
+    make_single_line_splitter,
+    Splitter,
+    Probe,
+)
 
-
-# -------------------- Generic splitting utils -------------------- #
-# TODO : merge with other splitting utils voir #391
-
-T1 = TypeVar("T1")
-T2 = TypeVar("T2")
-
-RawSplit = Tuple[List[T1], T2, List[T1]]
-"""
-Generic type alias representing a raw search & split operation on a list of elements.
-It is subscribed like so `RawSplit[ElementType, MatchType]`
-It represents a tuple `(before, match, after)` where:
-- `before` is of type `List[ElementType]` and represents a
-    list of elements before the match.
-- `match` is of type `MatchType` and represents the matched element.
-- `after` is of type `List[ElementType]` and represents a
-    list of elements after the match.
-"""
-
-Splitter = Callable[[List[T1]], RawSplit[T1, T2] | None]
-"""
-Generic type alias for a function that takes a list of elements,
-and returns a single RawSplit result or None if no match is found.
-It is subscribed like so `Splitter[ElementType, MatchType]`
-"""
-
-
-@dataclass(frozen=True)
-class SplitMatch(Generic[T1]):
-    value: T1
-
-
-@dataclass(frozen=True)
-class SplitNotAMatch(Generic[T1]):
-    value: T1
-
-
-SplittedElement = SplitNotAMatch[List[T1]] | SplitMatch[T2]
-"""
-Generic type alias for an element in a splitted list.
-
-It is subscribed like so `SplittedElement[ElementType, MatchType]`
-It represents either:
-- `SplitNotAMatch[List[ElementType]]` which encapsulates a list of elements that did not match.
-- `SplitMatch[MatchType]` which contains the matched element.
-
-This is useful for processing a list of elements and tagging as matches or non-matches
-in a generic manner. This enables proper typing and handling of the results like so :
-
-```
-if isinstance(splitted_element, SplitMatch):
-    splitted_element.value  # This is of type MatchType
-elif isinstance(splitted_element, SplitNotAMatch):
-    splitted_element.value  # This is of type List[ElementType]
-```
-"""
-
-Probe = Callable[[List[T1], int], bool]
-
-
-@iter_func_to_list
-def split_elements(
-    elements: List[T1],
-    splitter: Splitter[T1, T2],
-) -> Iterator[SplittedElement[T1, T2]]:
-    """
-    Split a list of elements using the provided splitter function.
-
-    For example :
-
-    >>> some_numbers = [1, 3, 11, 10, 6, 23]
-    >>> def multiple_of_3(elements: List[int]) -> RawSplit[int, int] | None:
-    ...     for i, element in enumerate(elements):
-    ...         if element % 3 == 0:
-    ...             return elements[:i], element, elements[i + 1:]
-    ...     return None
-    >>> list(split_elements(some_numbers, multiple_of_3))
-    [SplitNotAMatch([1]), SplitMatch(3), SplitNotAMatch([11, 10]), SplitMatch(6), SplitNotAMatch([23])]
-    """  # noqa: E501
-    # Here we make a copy, because we don't know
-    # if the splitter will modify the elements.
-    elements = list(elements)
-    while elements:
-        result = splitter(elements)
-        if result is None:
-            yield SplitNotAMatch(elements)
-            break
-        before, match, elements = result
-
-        if before:
-            yield SplitNotAMatch(before)
-        yield SplitMatch(match)
-
-
-@iter_func_to_list
-def map_splitted_elements(
-    splitted_list: List[SplittedElement[T1, T2]],
-    map_func: Callable[[T2], T1],
-) -> Iterator[T1]:
-    """
-    Map a function over a list of SplittedElement.
-
-    For example :
-
-    >>> splitted_list = [
-    ...     SplitNotAMatch([1, 2]),
-    ...     SplitMatch('hello'),
-    ...     SplitNotAMatch([4]),
-    ...     SplitMatch('world'),
-    ... ]
-    >>> def map_func(word: str) -> str:
-    ...     return word.upper()
-    >>> list(map_splitted_elements(splitted_list, map_func))
-    [1, 2, 'HELLO', 4, 'WORLD']
-    """
-    for splitted_element in splitted_list:
-        if isinstance(splitted_element, SplitMatch):
-            yield map_func(splitted_element.value)
-        else:
-            yield from splitted_element.value
-
-
-@iter_func_to_list
-def flat_map_splitted_elements(
-    splitted_list: List[SplittedElement[T1, T2]],
-    map_func: Callable[[T2], List[T1]],
-) -> Iterator[T1]:
-    for splitted_element in splitted_list:
-        if isinstance(splitted_element, SplitMatch):
-            yield from map_func(splitted_element.value)
-        else:
-            yield from splitted_element.value
-
-
-def split_before_match(
-    elements: List[T1],
-    is_matching: Probe[T1],
-) -> Tuple[List[T1], List[T1]]:
-    """
-    Split the input list into two parts, by using the `is_matching` function.
-
-    Examples :
-
-    strings = ["a", "b", "c"]
-    >>> split_before_match(strings, lambda s: s == "b")
-    (["a"], ["b", "c"])
-    >>> split_before_match(strings, lambda s: s == "d")
-    (["a", "b", "c"], [])
-    >>> split_before_match(strings, lambda s: s == "a")
-    ([], ["a", "b", "c"])
-    """
-    i = 0
-    while i < len(elements):
-        if is_matching(elements, i):
-            break
-        i += 1
-    return elements[:i], elements[i:]
-
-
-def make_single_line_splitter(
-    is_matching: Probe[T1],
-) -> Splitter:
-    """
-    Splits around the first matching element.
-
-    For example :
-
-    >>> strings = ["a", "b", "b", "c"]
-    >>> splitter = make_single_line_splitter(lambda s: s == "b")
-    >>> splitter(strings)
-    (["a"], ["b"], ["b", "c"])
-    """
-
-    def _splitter(elements: List[T1]) -> RawSplit[T1, List[T1]] | None:
-        before, after = split_before_match(elements, is_matching)
-        if after:
-            return (before, [after[0]], after[1:])
-        return None
-
-    return _splitter
-
-
-def make_while_splitter(
-    start_condition: Probe[T1],
-    while_condition: Probe[T1],
-) -> Splitter[T1, List[T1]]:
-    """
-    Starts the split at the first element matched by `start_condition`, and continues
-    to match until the first element that does not match `while_condition`.
-
-    For example :
-
-    >>> strings = ["a", "b", "b", "c"]
-    >>> splitter = make_while_splitter(lambda s, i: s == "b", lambda s, i: s == "b")
-    >>> splitter(strings)
-    (["a"], ["b", "b"], ["c"])
-    """
-
-    def _splitter(elements: List[T1]) -> RawSplit[T1, List[T1]] | None:
-        before, after = split_before_match(elements, start_condition)
-        if not after:
-            return None
-        match, after = split_before_match(
-            after, lambda elements, index: not while_condition(elements, index)
-        )
-        return before, match, after
-
-    return _splitter
-
-
-def negate(
-    probe: Probe[T1],
-) -> Probe[T1]:
-    """
-    Negates a probe function.
-
-    For example :
-
-    >>> strings = ["a", "b"]
-    >>> is_b = lambda elements, index: elements[index] == "b"
-    >>> is_not_b = negate(is_b)
-    >>> is_not_b(strings, 0)
-    True
-    >>> is_not_b(strings, 1)
-    False
-    """
-
-    def _negated_probe(elements: List[T1], index: int) -> bool:
-        return not probe(elements, index)
-
-    return _negated_probe
-
-
-# -------------------- Segmentation step splitting utils -------------------- #
 
 NodeOrText = Union[TextSegment, "Node"]
 
@@ -285,28 +51,6 @@ class Node:
     type: str
     children: List[NodeOrText]
     data: Dict[str, Any] = field(default_factory=dict)
-
-
-def group_text_segments_splitter(
-    elements: List[NodeOrText],
-) -> RawSplit[NodeOrText, List[TextSegment]] | None:
-    """
-    Splitter to enable grouping of TextSegment elements.
-    """
-    elements = list(elements)
-    before: List[NodeOrText] = []
-    match: List[TextSegment] = []
-    while elements and is_node(elements[0]):
-        before.append(elements[0])
-        elements.pop(0)
-
-    while elements and isinstance(elements[0], TextSegment):
-        match.append(elements[0])
-        elements.pop(0)
-
-    if match:
-        return (before, match, elements)
-    return None
 
 
 def pick_if_inline_node_followed_by_match(
@@ -348,7 +92,7 @@ def pick_if_inline_node_followed_by_match(
     return _pick_inline_nodes_probe
 
 
-def reject_if_not_text_segment(
+def pick_text_segment(
     probe: Probe[NodeOrText],
 ) -> Probe[NodeOrText]:
     def _text_segment_probe(elements: List[NodeOrText], index: int) -> bool:
@@ -391,8 +135,8 @@ def make_while_splitter_for_text_segments(
     while_condition: Probe[NodeOrText],
 ) -> Splitter[NodeOrText, List[NodeOrText]]:
     return make_while_splitter(
-        reject_if_not_text_segment(start_condition),
-        pick_if_inline_node_followed_by_match(reject_if_not_text_segment(while_condition)),
+        pick_text_segment(start_condition),
+        pick_if_inline_node_followed_by_match(pick_text_segment(while_condition)),
     )
 
 
@@ -400,8 +144,20 @@ def make_single_line_splitter_for_text_segments(
     is_matching: Probe[NodeOrText],
 ) -> Splitter[NodeOrText, List[NodeOrText]]:
     return make_single_line_splitter(
-        is_matching=reject_if_not_text_segment(is_matching),
+        is_matching=pick_text_segment(is_matching),
     )
+
+
+group_text_segments_splitter = cast(
+    Splitter[NodeOrText, List[TextSegment]],
+    make_while_splitter(
+        pick_text_segment(lambda elements, index: True),
+        pick_if_inline_node_followed_by_match(pick_text_segment(lambda elements, index: True)),
+    ),
+)
+"""
+Splitter to enable grouping of TextSegment elements.
+"""
 
 
 def is_node(node: Node | TextSegment, type_in: List[str] | None = None) -> TypeGuard[Node]:
