--- arretify/utils/merge.py
@@ -1,49 +1,0 @@
-#
-# Copyright (c) 2025 Direction générale de la prévention des risques (DGPR).
-#
-# This file is part of Arrêtify.
-# See https://github.com/mte-dgpr/arretify for further info.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-from typing import Iterator, Iterable
-
-from arretify.types import PageElementOrString
-
-
-def merge_strings(
-    str_gen: Iterable[str],
-) -> str:
-    return "".join(str_gen)
-
-
-def merge_strings_ignore_page_elements(
-    str_or_element_gen: Iterable[PageElementOrString],
-) -> Iterator[PageElementOrString]:
-    """
-    Example:
-        >>> elements = ["Hello-", "world!", Tag(name="p"), "More text"]
-        >>> list(merge_strings_ignore_page_elements(elements))
-        ['Hello-world!', <p></p>, 'More text']
-    """
-    accumulator: str | None = None
-    for str_or_element in str_or_element_gen:
-        if isinstance(str_or_element, str):
-            accumulator = (accumulator or "") + str_or_element
-        else:
-            if accumulator is not None:
-                yield accumulator
-            accumulator = None
-            yield str_or_element
-    if accumulator is not None:
-        yield accumulator
